requirements.txt
django-widget-tweaks
django-dynamic-raw-id